use +Extensions filename in new extensions
